# docs/swagger.yaml
swagger: "2.0"
info:
  description: "### *Simple Postman documentation*"
  version: "1.0"
  title: "godel"
  contact: {}
host: "127.0.0.1:3000"
basePath: "/api"
tags:
  - name: "Posts"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /posts:
    get:
      tags:
        - "Posts"
      summary: "Get All Posts"
      description: "## Get a list of all existing records in the database."
      operationId: "GetAllPosts"
      produces:
        - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          examples:
            application/json:
              msg: "Posts found"
              result:
                - post_id: "93faa45c-9c73-4ecb-baba-4809097d34f3"
                  title: "Mirazh"
                  description: "Postdsf description"
                  price: 11.9213000000000004519051799434237182140350341796875
                  images:
                    - image_url: "1233"
                    - image_url: "1233"
                    - image_url: "1233"
                  created_at: "2021-10-06T18:18:25.652Z"
                - post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
                  title: "Some title"
                  description: "Post description"
                  price: 11.9213000000000004519051799434237182140350341796875
                  images:
                    - image_url: "https://1233.com/image"
                    - image_url: "https://12333.com/image"
                    - image_url: "https://12334.com/image"
                  created_at: "2021-10-06T19:28:48.798Z"
              status: "success"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:29:13 GMT"
            Content-Length:
              type: "string"
              default: "592"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Success"
        "409":
          description: "Conflict"
          examples:
            application/json:
              msg: "No posts yet"
              result: ""
              status: "error"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 20:00:10 GMT"
            Content-Length:
              type: "string"
              default: "53"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Noposts1"
      deprecated: false
    post:
      tags:
        - "Posts"
      summary: "New Post"
      description: "## Create a new record."
      operationId: "NewPost"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "Body"
          description: ""
          required: true
          schema:
            $ref: "#/definitions/NewPostRequest"
          x-exportParamName: "Body"
      responses:
        "200":
          description: "OK"
          examples:
            application/json:
              msg: "Created post"
              result:
                post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
              status: "success"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:28:47 GMT"
            Content-Length:
              type: "string"
              default: "101"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Success1"
        "400":
          description: "Bad Request"
          examples:
            application/json:
              msg: "Validation error"
              result:
                - field: "Post.Images"
                  tag: "max"
                  value: "3"
              status: "error"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:59:14 GMT"
            Content-Length:
              type: "string"
              default: "102"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Validationerror1"
        "500":
          description: "Internal Server Error"
          examples:
            application/json:
              msg: "Incorrect input"
              result:
                Value: "\"\\\"11.9213\\\",\\n    \\\"images\\\": [\\n     ...\""
                Type: {}
                Offset: 0
                Struct: "models.Post"
                Field: "price"
              status: "error"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:58:33 GMT"
            Content-Length:
              type: "string"
              default: "179"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Incorrectinput1"
      deprecated: false
  /posts/{postId}:
    get:
      tags:
        - "Posts"
      summary: "Get Single Post"
      description: "## Get a single record by id."
      operationId: "GetSinglePost"
      produces:
        - "application/json"
      parameters:
        - name: "postId"
          in: "path"
          description: "ID of post to return"
          required: true
          type: "string"
          x-exportParamName: "PostId"
      responses:
        "200":
          description: "OK"
          examples:
            application/json:
              message: "Post found"
              result:
                image:
                  image_url: "https://1233.com/image"
                price: 11.9213000000000004519051799434237182140350341796875
                title: "Some title"
              status: "success"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:31:14 GMT"
            Content-Length:
              type: "string"
              default: "138"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Success2"
        "404":
          description: "Not Found"
          examples:
            application/json:
              message: "No post present"
              result: ""
              status: "error"
          headers:
            Date:
              type: "string"
              default: "Wed, 06 Oct 2021 19:57:25 GMT"
            Content-Length:
              type: "string"
              default: "60"
            Vary:
              type: "string"
              default: "Origin"
            Access-Control-Allow-Origin:
              type: "string"
              default: "*"
          schema:
            $ref: "#/definitions/Notfound1"
      deprecated: false
definitions:
  Success:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        type: "array"
        items:
          $ref: "#/definitions/Result"
      status:
        type: "string"
    title: "Success"
    example:
      msg: "Posts found"
      result:
        - post_id: "93faa45c-9c73-4ecb-baba-4809097d34f3"
          title: "Mirazh"
          description: "Postdsf description"
          price: 11.9213000000000004519051799434237182140350341796875
          images:
            - image_url: "1233"
            - image_url: "1233"
            - image_url: "1233"
          created_at: "2021-10-06T18:18:25.652Z"
        - post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
          title: "Some title"
          description: "Post description"
          price: 11.9213000000000004519051799434237182140350341796875
          images:
            - image_url: "https://1233.com/image"
            - image_url: "https://12333.com/image"
            - image_url: "https://12334.com/image"
          created_at: "2021-10-06T19:28:48.798Z"
      status: "success"
  Result:
    type: "object"
    required:
      - "created_at"
      - "description"
      - "images"
      - "post_id"
      - "price"
      - "title"
    properties:
      post_id:
        type: "string"
      title:
        type: "string"
      description:
        type: "string"
      price:
        type: "number"
        format: "double"
      images:
        type: "array"
        items:
          $ref: "#/definitions/Image"
      created_at:
        type: "string"
    title: "Result"
    example:
      post_id: "93faa45c-9c73-4ecb-baba-4809097d34f3"
      title: "Mirazh"
      description: "Postdsf description"
      price: 11.9213000000000004519051799434237182140350341796875
      images:
        - image_url: "1233"
        - image_url: "1233"
        - image_url: "1233"
      created_at: "2021-10-06T18:18:25.652Z"
  Image:
    type: "object"
    required:
      - "image_url"
    properties:
      image_url:
        type: "string"
    title: "Image"
    example:
      image_url: "1233"
  Withsorting:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        type: "array"
        items:
          $ref: "#/definitions/Result"
      status:
        type: "string"
    title: "Withsorting"
    example:
      msg: "Posts found"
      result:
        - post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
          title: "Some title"
          description: "Post description"
          price: 11.9213000000000004519051799434237182140350341796875
          images:
            - image_url: "https://1233.com/image"
            - image_url: "https://12333.com/image"
            - image_url: "https://12334.com/image"
          created_at: "2021-10-06T19:28:48.798Z"
        - post_id: "93faa45c-9c73-4ecb-baba-4809097d34f3"
          title: "Mirazh"
          description: "Postdsf description"
          price: 11.9213000000000004519051799434237182140350341796875
          images:
            - image_url: "1233"
            - image_url: "1233"
            - image_url: "1233"
          created_at: "2021-10-06T18:18:25.652Z"
      status: "success"
  Noposts1:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        type: "string"
      status:
        type: "string"
    title: "Noposts1"
    example:
      msg: "No posts yet"
      result: ""
      status: "error"
  NewPostRequest:
    type: "object"
    required:
      - "description"
      - "images"
      - "price"
      - "title"
    properties:
      title:
        type: "string"
      description:
        type: "string"
      price:
        type: "number"
        format: "double"
      images:
        type: "array"
        items:
          $ref: "#/definitions/Image"
    title: "NewPostRequest"
    example:
      title: "Some title"
      description: "Post description"
      price: 11.9213000000000004519051799434237182140350341796875
      images:
        - image_url: "https://1233.com/image"
        - image_url: "https://12333.com/image"
        - image_url: "https://12334.com/image"
  Success1:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        $ref: "#/definitions/Result2"
      status:
        type: "string"
    title: "Success1"
    example:
      msg: "Created post"
      result:
        post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
      status: "success"
  Result2:
    type: "object"
    required:
      - "post_id"
    properties:
      post_id:
        type: "string"
    title: "Result2"
    example:
      post_id: "2acc839d-20fe-4020-8803-f4f9901d1c5d"
  Result3:
    type: "object"
    required:
      - "Field"
      - "Offset"
      - "Struct"
      - "Type"
      - "Value"
    properties:
      Value:
        type: "string"
      Type:
        type: "object"
        properties: {}
      Offset:
        type: "integer"
        format: "int32"
      Struct:
        type: "string"
      Field:
        type: "string"
    title: "Result3"
    example:
      Value: "\"\\\"11.9213\\\",\\n    \\\"images\\\": [\\n     ...\""
      Type: {}
      Offset: 0
      Struct: "models.Post"
      Field: "price"
  Incorrectinput1:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        $ref: "#/definitions/Result3"
      status:
        type: "string"
    title: "Incorrectinput1"
    example:
      msg: "Incorrect input"
      result:
        Value: "\"\\\"11.9213\\\",\\n    \\\"images\\\": [\\n     ...\""
        Type: {}
        Offset: 0
        Struct: "models.Post"
        Field: "price"
      status: "error"
  Result5:
    type: "object"
    required:
      - "field"
      - "tag"
      - "value"
    properties:
      field:
        type: "string"
      tag:
        type: "string"
      value:
        type: "string"
    title: "Result5"
    example:
      field: "Post.Images"
      tag: "max"
      value: "3"
  Validationerror1:
    type: "object"
    required:
      - "msg"
      - "result"
      - "status"
    properties:
      msg:
        type: "string"
      result:
        type: "array"
        items:
          $ref: "#/definitions/Result5"
      status:
        type: "string"
    title: "Validationerror1"
    example:
      msg: "Validation error"
      result:
        - field: "Post.Images"
          tag: "max"
          value: "3"
      status: "error"
  Success2:
    type: "object"
    required:
      - "message"
      - "result"
      - "status"
    properties:
      message:
        type: "string"
      result:
        $ref: "#/definitions/Result7"
      status:
        type: "string"
    title: "Success2"
    example:
      message: "Post found"
      result:
        image:
          image_url: "https://1233.com/image"
        price: 11.9213000000000004519051799434237182140350341796875
        title: "Some title"
      status: "success"
  Result7:
    type: "object"
    required:
      - "image"
      - "price"
      - "title"
    properties:
      image:
        $ref: "#/definitions/Image"
      price:
        type: "number"
        format: "double"
      title:
        type: "string"
    title: "Result7"
    example:
      image:
        image_url: "https://1233.com/image"
      price: 11.9213000000000004519051799434237182140350341796875
      title: "Some title"
  Withmorefields:
    type: "object"
    required:
      - "message"
      - "result"
      - "status"
    properties:
      message:
        type: "string"
      result:
        $ref: "#/definitions/Result8"
      status:
        type: "string"
    title: "Withmorefields"
    example:
      message: "Post found"
      result:
        description: "Post description"
        images:
          - image_url: "https://1233.com/image"
          - image_url: "https://12333.com/image"
          - image_url: "https://12334.com/image"
        price: 11.9213000000000004519051799434237182140350341796875
        title: "Some title"
      status: "success"
  Result8:
    type: "object"
    required:
      - "description"
      - "images"
      - "price"
      - "title"
    properties:
      description:
        type: "string"
      images:
        type: "array"
        items:
          $ref: "#/definitions/Image"
      price:
        type: "number"
        format: "double"
      title:
        type: "string"
    title: "Result8"
    example:
      description: "Post description"
      images:
        - image_url: "https://1233.com/image"
        - image_url: "https://12333.com/image"
        - image_url: "https://12334.com/image"
      price: 11.9213000000000004519051799434237182140350341796875
      title: "Some title"
  Notfound1:
    type: "object"
    required:
      - "message"
      - "result"
      - "status"
    properties:
      message:
        type: "string"
      result:
        type: "string"
      status:
        type: "string"
    title: "Notfound1"
    example:
      message: "No post present"
      result: ""
      status: "error"